requirements.txt
acres==0.5.0
altair==6.0.0
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
blinker==1.9.0
cachetools==7.0.4
certifi==2026.2.25
charset-normalizer==3.4.5
ci-info==0.4.0
click==8.3.1
colorama==0.4.6
configobj==5.0.9
configparser==7.2.0
distro==1.9.0
dotenv==0.9.9
etelemetry==0.3.1
exceptiongroup==1.3.1
faiss-cpu==1.7.4
filelock==3.25.0
gitdb==4.0.12
GitPython==3.1.46
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
idna==3.11
importlib_resources==6.5.2
isodate==0.7.2
Jinja2==3.1.6
jiter==0.13.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
looseversion==1.3.0
lxml==6.0.2
MarkupSafe==3.0.3
narwhals==2.17.0
networkx==3.4.2
nibabel==5.4.0
nipype==1.11.0
numpy==1.26.4
openai==2.26.0
packaging==26.0
pandas==2.3.3
pathlib==1.0.1
pillow==12.1.1
protobuf==6.33.5
prov==2.1.1
puremagic==1.30
pyarrow==23.0.1
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
pydot==4.0.1
PyMuPDF==1.23.26
PyMuPDFb==1.23.22
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2026.1.post1
pyxnat==1.6.4
rdflib==7.6.0
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
scipy==1.15.3
simplejson==3.20.2
six==1.17.0
smmap==5.0.3
sniffio==1.3.1
streamlit==1.55.0
tenacity==9.1.4
toml==0.10.2
tornado==6.5.4
tqdm==4.67.3
traits==7.1.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
watchdog==6.0.0
